requirements.txt
flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
openai==0.27.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-JWT-Extended==4.3.1
stripe==2.60.0
bcrypt==3.2.0
email-validator==1.1.3
requests==2.31.0
PyMuPDF==1.21.1
python-docx==0.8.11 